- more new nice stuff;

+++ core/src/com/krld/donttouch/game/model/Player.java
@@ -3,8 +3,8 @@ package com.krld.donttouch.game.model;
 import com.badlogic.gdx.graphics.Color;
 
 public class Player implements Coordinates {
-    private int x;
-    private int y;
+    private float x;
+    private float y;
     private float radius;
     private Color color;
 
@@ -41,11 +41,11 @@ public class Player implements Coordinates {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
 }
